pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "magi-system2"
version = "0.1.0"
description = "Multi-persona AI discussion system with dynamic persona generation and adaptive facilitation"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.67.0",
    "pydantic>=2.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.0",
    "websockets>=15.0",
]

[project.scripts]
magi2 = "magi_system2.cli:main"

[project.urls]
Repository = "https://github.com/nlink-jp/magi-system2"

[tool.setuptools.packages.find]
include = ["magi_system2*"]

[tool.setuptools.package-data]
magi_system2 = ["templates/*.html", "static/*.css", "static/*.js"]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=8.0.0",
]
